:cat: Change(directory): core-is-outside-src
