+++ core/config.ts
