+++ core/errors.ts
