+++ core/guards.ts
