+++ core/interfaces/error.interface.ts
