+++ core/interfaces/index.ts
