+++ core/interfaces/response.interface.ts
