+++ package.json
@@ -23,8 +23,6 @@
     "cross-env": "^7.0.3",
     "dotenv": "^16.1.4",
     "express": "^4.18.2",
-    "express-async-errors": "^3.1.1",
-    "express-async-handler": "^1.2.0",
     "tsconfig-paths": "^4.2.0",
     "zod": "^3.21.4"
   },
