+++ src/routes.ts
@@ -1,10 +1,5 @@
-import { Controller, IRoute } from "@core/interfaces";
+import { IRoute } from "@core/interfaces";
 import { hello } from "@app/hello";
-
-// sample welcome route for (root) /
-const welcome: Controller = async (_, res) => {
-  res.toJson({ message: "exTS 🐈! custom backend framework." });
-};
 
 /**==========================================================*
  * @exTS custom express.js modified framework
@@ -12,12 +7,7 @@ const welcome: Controller = async (_, res) => {
  *===========================================================*/
 
 const routes: IRoute[] = [
-  ...hello, // hello application hooked
-  {
-    endpoint: "/",
-    controller: welcome,
-    method: "get",
-  },
+  ...hello, // hello application added
 ];
 
 export default routes;
